+++ src/components/index.ts
@@ -1,2 +1,2 @@
 // Common Components:
-export  { default as LoadingComponent} from './common/loading/Loading';
+export  { default as LoadingComponent} from "./common/loading/Loading";
