+++ src/components/pipeline/PipelineTree.tsx
@@ -1,9 +1,9 @@
 import React, { createRef, RefObject } from "react";
 import * as d3 from "d3";
 import * as cola from "webcola";
-import TreeModel, { ITreeChart } from "../../../api/models/tree.model";
-import TreeNodeModel, { INode } from "../../../api/models/tree-node.model";
-import { IPluginItem } from "../../../api/models/pluginInstance.model";
+import TreeModel, { ITreeChart } from "../../api/models/tree.model";
+import TreeNodeModel, { INode } from "../../api/models/tree-node.model";
+import { IPluginItem } from "../../api/models/pluginInstance.model";
 import * as _ from "lodash";
 interface ITreeProps {
   selected: IPluginItem;
