+++ src/pages/Feeds/feed.scss
@@ -5,7 +5,7 @@
 */
 
 
-@import '../../../assets/scss/helpers/variables'; 
+@import '../../assets/scss/helpers/variables'; 
 $label-font-size: 0.8125rem;
 $data-font-size: 0.875rem;
 .feed-info-block{
